requirements.txt
Django
djangorestframework
psycopg2

#unittests
pytest
pytest-django
factory-boy
